Trying Problem 23 again. Needs testing
This is one of a few approaches that I'm thinking about approaching
this problem with.
The program actively seeks out the next abundant number, but I
suspect that it spends a few too many clock cycles and too much
complexity to be of any use.

Another approach that I may try next is to find all abundant numbers,
calculate all sums of each pair, then adding the numbers which work.
Watch out for numbers which may be calculated twice.